P-131 fix dashboard suite à changement coté serveur

+++ clientLeger/app/src/main/java/com/projet/clientleger/data/api/model/account/GameHistory.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 data class GameHistory(val games:ArrayList<Game>,
                        val nbGamePlayed:String,
                        val winPercentage:Float,
-                       val averageGameTime:Long,
+                       val averageGameTime:Float,
                        val totalTimePlayed:Long,
                        val bestScoreSolo:Int,
                        val bestScoreCoop:Int)
